app/requirements.txt
fastapi==0.115.5
uvicorn==0.30.3
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
redis==5.0.4
pyJWT==2.10.0
werkzeug==3.1.3
httpx==0.22.0
pycountry==24.6.1